+++ database/migrations/2019_05_07_124632_create_sprint_table.php
@@ -16,7 +16,7 @@ class CreateSprintTable extends Migration
         Schema::create('sprint', function (Blueprint $table) {
             $table->bigIncrements('sprint_id');
             $table->dateTime('start_date');
-            $table->bigInteger('sprint_info_id');
+            $table->integer('sprint_info_id');
             $table->timestamps();
         });
     }
